시큐리티, id, pw 찾기 에러, 결재문서조회 에러 해결

+++ arion/src/main/java/com/arion/app/home/main/web/HomeController.java
@@ -133,7 +133,7 @@ public class HomeController {
     
     @PostMapping("/home/findId")
     @ResponseBody
-    public Map<String, String> findCompanyId(@RequestParam("ceoName") String ceoName, @RequestParam("companyBusinessNumber") int companyBusinessNumber, Model model) {
+    public Map<String, String> findCompanyId(@RequestParam("ceoName") String ceoName, @RequestParam("companyBusinessNumber") long companyBusinessNumber, Model model) {
     	CompanyVO companyVO = csvc.findId(ceoName, companyBusinessNumber);
     	Map<String, String> response = new HashMap<>();
 		if(companyVO != null) {
