+++ arion/src/main/java/com/arion/app/security/mapper/UserMapper.java
@@ -17,7 +17,7 @@ public interface UserMapper {
     int checkId(@Param("companyId") String companyId);
     String selectCompanyName(@Param("companyCode") String companyCode);
     
-    CompanyVO findId (@Param("ceoName") String ceoName, @Param("companyBusinessNumber") int companyBusinessNumber);
+    CompanyVO findId (@Param("ceoName") String ceoName, @Param("companyBusinessNumber") long companyBusinessNumber);
     int updatePw (@Param("companyCode") String companyCode, @Param("companyId") String companyId, @Param("companyPw") String companyPw);
     String selectEmail (@Param("companyCode") String companyCode, @Param("companyId") String companyId);
